Separate centralized and decentralized optimizers
